backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pymongo==4.6.1
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.32.3
pydantic==1.10.15
email-validator==2.1.0.post1
stripe==8.10.0
